refactor: move AmountInput component to UI dir
