--- app/components/AmountInput/index.js
@@ -1,3 +1,0 @@
-import AmountInput from './AmountInput'
-
-export default AmountInput
