+++ app/components/Contacts/SubmitChannelForm/SubmitChannelForm.js
@@ -1,8 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import FaExclamationCircle from 'react-icons/lib/fa/exclamation-circle'
-import AmountInput from 'components/AmountInput'
-import { Button, Dropdown } from 'components/UI'
+import { AmountInput, Button, Dropdown } from 'components/UI'
 import { FormattedNumber, FormattedMessage } from 'react-intl'
 import messages from './messages'
 import styles from './SubmitChannelForm.scss'
